modellemelerde değişiklikler yapıldı.

+++ MadrinHotelCRM.Entities/Models/Tarife.cs
@@ -13,7 +13,7 @@ namespace MadrinHotelCRM.Entities.Models
         public decimal Fiyat { get; set; }
         public DateTime BaslangicTarihi { get; set; }
         public DateTime BitisTarihi { get; set; }
-        public decimal IndirimOrani { get; set; }
+        public int IndirimOrani { get; set; }
         public ICollection<OdaTarife> OdaTarifeleri { get; set; }// OdaTarife ile bire çok ilişki
     }
 }
